Desktop/python-projects/python-projects/myproject/requirement.txt
asgiref==3.8.1
cachetools==5.5.1
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
Django==5.1.5
django-ckeditor-5==0.2.15
django-js-asset==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.1
google-cloud-translate==3.19.0
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
legacy-cgi==2.6.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
redis==5.2.1
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.3
urllib3==2.3.0